Rename `BloodPressureHistoryListItem` to `BloodPressureHistoryListItem_Old`
