+++ app/src/androidTest/java/org/simple/clinic/di/TestAppComponent.kt
@@ -16,7 +16,7 @@ import org.simple.clinic.benchmark.patientregistration.PatientRegistrationBenchm
 import org.simple.clinic.benchmark.recentpatient.RecentPatientsBenchmark
 import org.simple.clinic.bloodsugar.BloodSugarHistoryListItemDataSourceTest
 import org.simple.clinic.bloodsugar.BloodSugarRepositoryAndroidTest
-import org.simple.clinic.bp.BloodPressureHistoryListItemDataSourceTest
+import org.simple.clinic.bp.BloodPressureHistoryListItemOldDataSourceTest
 import org.simple.clinic.bp.BloodPressureRepositoryAndroidTest
 import org.simple.clinic.drugs.PrescriptionRepositoryAndroidTest
 import org.simple.clinic.drugs.search.DrugRepositoryAndroidTest
@@ -118,7 +118,7 @@ interface TestAppComponent {
   fun inject(target: Migration57AndroidTest)
   fun inject(target: Migration58AndroidTest)
   fun inject(target: Migration59AndroidTest)
-  fun inject(target: BloodPressureHistoryListItemDataSourceTest)
+  fun inject(target: BloodPressureHistoryListItemOldDataSourceTest)
   fun inject(target: BloodSugarHistoryListItemDataSourceTest)
   fun inject(target: TeleconsultationFacilityWithMedicalOfficersTest)
   fun inject(target: TeleconsultationFacilityRepositoryAndroidTest)
